Drive Publisher Logic to Base Publisher.
Borrowing a little from the world of Rails here, create a base publisher
that all other publishers can inherent from. This allows the monkeying
that's required with Dry Events Publishers to be a bit more hidden and
type-able.

Without this Sorbet will complain about including Dry Events Publisher
which is a class and not a module... Rather it's a class masquerading as
a module. To get around this, build a publisher method that takes an id
and creates the publisher class for the application.