Express Return Of Dry Publisher Registry
Since the application is adding to the registry now, not relying on Dry
Events, the registry and Concurrent Map needs type sigs. This adds those
required sigs.

The use of void is likely incorrect here. It's likely some hash like
object. However, not being sure about the proper sig, leaving this as
void for now.