Add A Tasks Definition File.
Much like VSCode tasks, these tasks interact with files, buffers, etc to
be able 'do' things with the application. Currently this is setup to use
prettier on the open file and run sorbet for type checking problems that
we might have introduced.

Yes, there is ALE which is great if you want to HOUSE your batteries.
However, running ALE plus a language server causes a lot of micro cuts
when it comes to battery performance. Keep to linting and type checking
on demand as well as running the build in Neovim LSP for language server
needs