Test Product Get Route.
The route should return the result of a product. It's an expectation
that this route will be called subsequently after the creation of a
product.

This also introduces a very static version of a product. This will
likely change. Wiring something like Stripe up to the application to
gain access to storing products there, rather than locally in a database
or in a constant (current implementation).