HAL JSON Validations (#1)
* Install Dry Validations.

This will be the main item that checks the input params from PUTs/POSTs
throughout the application. It will also create contracts that are used
in the testing suite to validate that the responses align to the HAL
JSON spec.

* Fix File Path For Tasks.

Filename is only going to provide the file name and not the full path to
the file that needs to be fixed/typed. Fix this by providing the
filepath.

* Define A Simple Body Schema And Middleware.

Build a way for the API to seamlessly deal with bad body params using
Dry Validators and Rack middleware. This gives a simple way to make all
API POST/PUT routes abide by a structured input.

* Test App Idempotent.

Make sure to only spin up a single app instance for the entire (needed)
test suite.

* Install Ruby Ulid Library.

Since the application deals with events, there needs to be some way to
uniquely identify these events. Reach for ULID's due to them having
almost no duplicates if many events were to fire within the same
millisecond.

* Wire ULID's to ProductCreate Route.

Now the event system has something to hinge on when sending information
back to the user. When someone creates a product, they can use the
subsequent Location header as a reference to the newly created product.

* Product Publisher Emits Requested Events.

Allow the product publisher to emit events that deal with a consumer
requesting a specific product. This will ultimately flow down through to
when a product is requested at product/id

* Create A Base Hal Params Object.

Create a HalBase param schema to mix in with contracts that test the
return from all of the API based endpoints. This also isn't all of the
HAL output spec, just a good amount to start.

* Update Hanami And Rack Test.

Move back to the world of rack-test to make full stack request tests easier.
Including a bump to make nested middleware work properly in Hanami API.
This bump changes how the app is bootstrapped in the config.ru file.

* Pull In Rack Test To Spec Helper.

Utilize Rack's Builder pattern to grab the first application in the config.ru.
Use the Rack Test methods to allow tests to have easy access to the running
app (just for request specs though).

* Utilize Rack Test Methods In Request Specs.

With the declaration of app method in the request helper,
there is no longer a need to call an HTTP method on an app instance.
Rack Test magic!

* Add Repositories To MiniTest Autoload.

This new group of files need to be added to the Minitest autoload.
Removing the now unused events file.

* Test Product Get Route.

The route should return the result of a product. It's an expectation
that this route will be called subsequently after the creation of a
product.

This also introduces a very static version of a product. This will
likely change. Wiring something like Stripe up to the application to
gain access to storing products there, rather than locally in a database
or in a constant (current implementation).

* Add Links To Repository Tests.

The repository itself will alter the product response for now. This adds
the HAL specific links part of the response.