Remove Extraneous Gems.
Why Validations when there is already Sorbet being used throughout the
project? Odd grab really and a mistake on my part. Remove this, dry
validations and mocha in favor of a much slimmer set of gems and to lean
more on Sorbet to define the interfaces/returns of objects (over dry
validations).