Add Links to Product Create Test.
Links should be a required item for the ingestion of application items.
This isn't specifically apart of the HAL specification but is touched on
in the HAL-FORMS documentation.

For the immediate future, rely on ingesting a type key for
put/patch/post requests and flush out what the internal structure of the
_links object should look like later