Rubyfmt for Formatting (#4)
* Prettier to Rubyfmt.

Give rubyfmt a bit of a spin now that there are pre-built binaries
available. Remove prettier since it won't be needed for formating and
wire up rubyfmt into the tasks file.

* Run Rubyfmt On Project.

Re align everything using this new formatting. Not much of a change but
definitely there is a favor towards double quotes everywhere and
sensible spacing in data structures/blocks