Stripe For Products (#5)
This is a seemingly HUGE pull request but, it really is just the Sorbet types to make the Stripe for products actually work. This pull request also fixes any outlying type issues and re-allows the application to successfully be built and type checked with GitHub Actions.

It also encompasses a move away from Hanami API and back to a router based system like Roda. While the Hanami API is probably powerful, the work here to get it working in a manner that was useful was way more difficult than 'simply' setting up a complex routing tree like Roda.

* Install Stripe and Stripe Mocks.

Product data will originate from Stripe. Originally this was going to be
done locally but, there really isn't a need since all payments will come
through Stripe (likely.)

* Ignore Envrc Files.

These items usually alter the environment for the specific folder.
Ignore these to be sensitive to secrets creeping into the repo.

* Wire Up Spec Helper For Stripe Mock.

Make these items a bit less stripe specific and wire them into the
test suite. This allows the application to use StripeMock without its'
knowledge bleeding into each test.

* Create Concept of Resources.

Resources will be a central space into the Stripe gem. It should allow
for an easy refactoring later on when and if the application needs to
  move to another service to manage the pricing/product information.

* Alter Product Repository For Stripe.

Change and green up tests around finding a single product by an ID. This
isn't all that great due to filtering through a list, especially when
Stripe has a single resource GET method around products. However,
incremental improvements and all.

* Go Back To A Defaultish Ruby Version File.

ASDF is way overkill, use docker

* Green Up Product Test.

This now utilizes StripeMock to mock a new product, retrieve it from the
server and pass it to the API.

* Begin Typing Resources From Stripe.

* Bail On Hanami API.

Move to Roda. Roda has a much smaller gem file footprint and since this
really is a testbed for Sorbet, move to something easier to type.
Musterman in Hanami Web would boot a server when introspecting gems and
ultimately fail.

Let's see if Roda fairs better?

* Generalize Testing OF The Body Params Validator.

Make this less Roda/Hanami/Framework specific and more of a cross rack
middleware utility.

* Avoid Freezing Routes In Dev.

Make the dev experience here (when running the application) much more
nice.

* Green Up All Tests.

Alter how the application reacts to nested routes, parameters, etc. This
completes the move from Hanami API to Roda.

* RubyFmt The World.

Stick with the Rust big brother of all Ruby Formatters. This has been
run project wide so a whole lot of single '' to "" have been generated.
BIG DIFF!

* Add In Test Key.

Allow Stripe Mock to pass successfully.

* Sorbet Bump!

* Rebuild Sorbet File.

* Build Types For Middleware.

Type sigs for the middleware that scans for _links on post params.

* Fix Problems with Auto Fixed Middleware.

Revert to Rack specific items.

* Consolidate The Resources Module.

For now drive all of the logic into the ProductRepository for a Stripe
Product. Utilize the StripeObject super class to correctly flesh out the
types of our specific Stripe Product.

* Fix Product Type Def.

Utilize the Stripe Product object to define the shape of the return over
the custom implementation that's built here.

* Green Up All Tests.