Rebuild Objects At App Level.
Rebuild/rehydrate Stripe Objects at the router level. This is unlikely
the best place for these but they tend to be a bit of a bear to type.
Typing the hashes that hydrate them is ultimately less verbose and
easier to reuse in the long run? Not sold but, let's see.