Better Stripe Types/Mocks (#6)
This pull request moves a lot of testing over to the Stripe Mock container versus using the Stripe Mock Ruby library. This is a more 'recent' and more up-to-date way to run tests locally for the Stripe gem. 

Included is a host of contributing/readme/license changes. This better explains what's going on here.